requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asn1crypto==1.3.0
astroid==2.3.3
attrs==19.3.0
awscli==1.17.10
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.8.0
black==19.10b0
bleach==3.1.0
boto3==1.9.253
botocore==1.12.253
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
codecov==2.0.15
colorama==0.4.1
coverage==5.0.3
credstash==1.15.0+indigo
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
ec2-metadata==2.1.0
elasticsearch==7.5.1
entrypoints==0.3
Fiona==1.8.13
future==0.18.2
GeoAlchemy2==0.6.3
geopandas==0.6.3
html5lib==1.0.1
idna==2.8
ijson==2.5.1
imagesize==1.2.0
importlib-metadata==1.5.0
indigo.dsm-utils==0.0.16
ipython-genutils==0.2.0
isort==4.3.21
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter-core==4.6.1
jupyterlab==2.1.0
jupyterlab-pygments==0.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
munch==2.5.0
nbconflux==0.7.0
nbconvert==6.0.0a0
nbformat==5.0.4
numpy==1.18.1
oscrypto==1.2.0
packaging==20.1
pandas==0.25.3
pandocfilters==1.4.2
pathspec==0.7.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
pycryptodomex==3.9.4
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypd==1.1.0
pyproj==2.4.2.post1
pyrsistent==0.15.7
pytest==5.3.5
pytest-cov==2.8.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
regex==2020.1.8
requests==2.22.0
requests-aws4auth==0.9
rsa==3.4.2
Rtree==0.8.3
s3transfer==0.2.1
seaborn==0.10.0
Shapely==1.7.0
six==1.14.0
slacker==0.13.0
snowballstemmer==2.0.0
snowflake-connector-python==2.1.3
snowflake-sqlalchemy==1.1.18
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.13
tabulate==0.8.6
telluslabs==1.21.4
testpath==0.4.4
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
unify==0.5
untokenize==0.1.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
yapf==0.29.0
zipp==2.1.0
